trip_api/requirements.txt
django
djangorestframework
geopy